# smart-ground-hub/src/main/java/ch/jp/shooting/dto/CreateUserRequest.java
package ch.jp.shooting.dto;

import org.jspecify.annotations.NullMarked;

@NullMarked
public class CreateUserRequest {
    private String username;
    private String password;
    private String role;

    public CreateUserRequest() {}

    public CreateUserRequest(String username, String password, String role) {
        this.username = username;
        this.password = password;
        this.role = role;
    }

    public String getUsername() { return username; }
    public void setUsername(String username) { this.username = username; }

    public String getPassword() { return password; }
    public void setPassword(String password) { this.password = password; }

    public String getRole() { return role; }
    public void setRole(String role) { this.role = role; }
}