# smart-ground-hub/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SmartGround API
  version: 2.0.0
  description: >
    REST API for IoT device management with MQTT-based device discovery and
    command signaling. Includes endpoints to manage device types, register
    devices, and trigger config pushes to SmartBox firmware via MQTT.

    ## Device Management

    All device operations use the flat `/api/devices` endpoint. Devices are
    always registered on a SmartBox (network access point) but are managed
    as a flat list with optional filtering by SmartBox or Range.
    Use **POST /api/devices** to register and immediately trigger a config push
    to the SmartBox firmware via MQTT.

    ## Breaking changes from v2.x

    - `POST /api/smart-boxes/{id}/devices` removed — use `POST /api/devices`
    - `GET /api/smart-boxes/{id}/devices` removed — use `GET /api/devices?smartBoxId=X`
    - `POST/DELETE /api/ranges/{rangeId}/devices/{deviceId}` removed — use `POST/DELETE /api/devices/{id}/range`
    - `SmartBoxResponse.deviceCount` removed
servers:
  - url: http://localhost:8080
    description: JVM App
  - url: http://localhost:8081
    description: Native App
security:
  - bearerAuth: []
paths:
...
  /api/users:
    post:
      summary: Create a new user
      operationId: createUser
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
        '409':
          description: User already exists

  # ─────────────────────────────────────────────────────────────────────────────
  # DeviceType
  # ─────────────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CreateUserRequest:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          description: Username for the new user
        password:
          type: string
          description: Password for the new user
        role:
          type: string
          description: User role (ADMIN or SHOOTER)
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT